+++ src/parameters/graph.rs
@@ -1,18 +1,18 @@
 pub mod graph {
 
     use bevy::prelude::Resource;
-    const WIDTH:f32 = 500.0;
-    const HEIGHT:f32 = 500.0;
-    const POINT_SIZE:f32 = 2.5;
-    const PADDING:f32 = 20.0;
+    const WIDTH:f64 = 500.0;
+    const HEIGHT:f64 = 500.0;
+    const POINT_SIZE:f64 = 2.5;
+    const PADDING:f64 = 20.0;
     const NUMBER_TICKS: u8 = 5;
 
     #[derive(Resource)]
     pub struct Parameters {
-        pub width: f32,
-        pub height: f32,
-        pub points_size: f32,
-        pub padding: f32,
+        pub width: f64,
+        pub height: f64,
+        pub points_size: f64,
+        pub padding: f64,
         pub number_ticks: u8,
     }
